refactor(sidebar): split Sidebar into focused modules with type-safe architecture
- Extract primitives (ActiveIndicator, UserAvatar, MiniAppIcon, SidebarTabIcon)
- Split menu/docked/footer into layout-aware components (SidebarMenu, SidebarDocked, SidebarFooter)
- Extract useSidebarResize hook for resize logic isolation
- Add SidebarVisibleLayout type alias, removing Exclude<> repetition
- Move miniAppTabs into SidebarMenuItem, decoupling 'minapp' hardcode from presentational layer
- Convert OpenClawSidebarIcon to createLucideIcon for unified LucideIcon type
- Simplify SidebarUser (merge avatarSrc/avatarFallback into avatar, remove unused email)
- Remove unused closeable/pinned fields from SidebarBaseTab
- Tighten SidebarMiniApp.logo from string|object to string|CompoundIcon
- Rename isEmojiAvatar to isTextAvatar for semantic clarity
- Convert renderSidebarTabIcon function to SidebarTabIcon component
