+++ src/renderer/src/components/Icons/SVGIcon.tsx
@@ -1,4 +1,5 @@
 import { lightbulbVariants } from '@renderer/utils/motionVariants'
+import { createLucideIcon } from 'lucide-react'
 import { motion } from 'motion/react'
 import { type SVGProps, useId } from 'react'
 
@@ -317,31 +318,15 @@ export function OpenClawIcon(props: SVGProps<SVGSVGElement>) {
   )
 }
 
-export function OpenClawSidebarIcon(props: SVGProps<SVGSVGElement>) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      stroke-width="2"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-      {...props}>
-      <path d="M14.4 5.2a5.3 5.3 0 0 1 2.022-2.548L16.8 2.4" />
-      <path d="M14.56 8.933v1" />
-      <path d="m15.1 18.933.81 2.65" />
-      <path d="M18.56 8.433c.833.333 2 1 2 2" />
-      <path d="M5.56 8.433c-.833.333-2 1-2 2" />
-      <path d="m7.91 18.933-.81 2.65" />
-      <path d="M9.56 8.933v1" />
-      <path d="M9.6 5.2a5.3 5.3 0 0 0-2.022-2.548L7.2 2.4" />
-      <circle cx="12" cy="12" r="7.2" />
-    </svg>
-  )
-}
+export const OpenClawSidebarIcon = createLucideIcon('OpenClaw', [
+  ['path', { d: 'M8 4Q6 1 4 2' }],
+  ['path', { d: 'M16 4q2-3 4-2' }],
+  ['circle', { cx: '12', cy: '12', r: '9' }],
+  ['circle', { cx: '9', cy: '10', r: '1.5', fill: 'currentColor', stroke: 'none' }],
+  ['circle', { cx: '15', cy: '10', r: '1.5', fill: 'currentColor', stroke: 'none' }],
+  ['path', { d: 'M9 21v2' }],
+  ['path', { d: 'M15 21v2' }]
+])
 
 // https://code.visualstudio.com/brand
 export const VSCodeIcon = (props: SVGProps<SVGSVGElement>) => {
