+++ src/renderer/src/components/Sidebar/types.ts
@@ -1,33 +1,44 @@
-import type { ElementType } from 'react'
+import type { CompoundIcon } from '@cherrystudio/ui'
+import type { LucideIcon } from 'lucide-react'
+
+export interface SidebarMiniApp {
+  id: string
+  color?: string
+  url?: string
+  logo?: string | CompoundIcon
+}
+
+export interface SidebarMiniAppTab {
+  id: string
+  title: string
+  type: 'miniapp'
+  miniApp: SidebarMiniApp
+}
 
 export interface SidebarMenuItem {
   id: string
   label: string
-  icon: ElementType
+  icon: LucideIcon
+  miniAppTabs?: SidebarMiniAppTab[]
 }
 
-export interface SidebarTab {
+export interface SidebarRouteTab {
   id: string
   title: string
-  icon: ElementType
-  closeable?: boolean
-  pinned?: boolean
-  sidebarDocked?: boolean
-  menuItemId?: string
-  miniAppId?: string
-  miniAppColor?: string
-  miniAppInitial?: string
-  miniAppUrl?: string
-  miniAppLogoUrl?: string
-  miniAppLogo?: string | object
+  type: 'route'
+  icon: LucideIcon
+  sourceMenuItemId?: string
+  dockTarget?: 'sidebar'
 }
+
+export type SidebarTab = SidebarRouteTab | SidebarMiniAppTab
 
 export type SidebarLayout = 'hidden' | 'icon' | 'vertical-card' | 'full'
 
+export type SidebarVisibleLayout = Exclude<SidebarLayout, 'hidden'>
+
 export interface SidebarUser {
   name: string
-  email: string
-  initial?: string
-  avatarSrc?: string
+  avatar?: string
   onClick?: () => void
 }
